SQLITE result writer fixed and tested

+++ src/main/kotlin/Results.kt
@@ -1,5 +1,5 @@
 import kotlinx.serialization.Serializable
-import java.util.concurrent.atomic.AtomicLong
+import java.util.concurrent.atomic.AtomicInteger
 
 @Serializable
 data class TestMethodInfo(
@@ -10,7 +10,7 @@ data class TestMethodInfo(
     var isParametrised: Boolean,
     val classInfo: TestClassInfo,
     val sourceMethod: SourceMethodInfo?,
-    val id: Long = IdGenerator.testMethodCounter.getAndIncrement()
+    val id: Int = IdGenerator.testMethodCounter.getAndIncrement()
 )
 
 @Serializable
@@ -19,7 +19,7 @@ data class TestClassInfo(
     val projectInfo: ProjectInfo,
     val moduleInfo: ModuleInfo,
     val sourceClass: SourceClassInfo?,
-    val id: Long = IdGenerator.testClassCounter.getAndIncrement()
+    val id: Int = IdGenerator.testClassCounter.getAndIncrement()
 )
 
 @Serializable
@@ -27,35 +27,35 @@ data class SourceMethodInfo(
     val name: String,
     val body: String,
     val sourceClass: SourceClassInfo,
-    val id: Long = IdGenerator.sourceMethodCounter.getAndIncrement()
+    val id: Int = IdGenerator.sourceMethodCounter.getAndIncrement()
 )
 
 @Serializable
 data class SourceClassInfo(
     val name: String,
     val moduleInfo: ModuleInfo,
-    val id: Long = IdGenerator.sourceClassCounter.getAndIncrement()
+    val id: Int = IdGenerator.sourceClassCounter.getAndIncrement()
 )
 
 @Serializable
 data class ProjectInfo(
     val name: String,
     val buildSystem: BuildSystem,
-    val id: Long = IdGenerator.projectCounter.getAndIncrement(),
+    val id: Int = IdGenerator.projectCounter.getAndIncrement(),
 )
 
 @Serializable
 data class ModuleInfo(
     val name: String,
     val projectInfo: ProjectInfo,
-    val id: Long = IdGenerator.moduleCounter.getAndIncrement()
+    val id: Int = IdGenerator.moduleCounter.getAndIncrement()
 )
 
 object IdGenerator {
-    val projectCounter: AtomicLong = AtomicLong(1)
-    val moduleCounter: AtomicLong = AtomicLong(1)
-    val testMethodCounter: AtomicLong = AtomicLong(1)
-    val sourceMethodCounter: AtomicLong = AtomicLong(1)
-    val testClassCounter: AtomicLong = AtomicLong(1)
-    val sourceClassCounter: AtomicLong = AtomicLong(1)
+    val projectCounter: AtomicInteger = AtomicInteger(1)
+    val moduleCounter: AtomicInteger = AtomicInteger(1)
+    val testMethodCounter: AtomicInteger = AtomicInteger(1)
+    val sourceMethodCounter: AtomicInteger = AtomicInteger(1)
+    val testClassCounter: AtomicInteger = AtomicInteger(1)
+    val sourceClassCounter: AtomicInteger = AtomicInteger(1)
 }
